minions/requirements.txt
mlx-lm
openai
python-dotenv